move periodic partitioning to fullset converter only to separate implementations
