+++ src/lamp_py/ingestion/backfill/convert_gtfs_rt_fullset.py
@@ -17,13 +17,13 @@ from lamp_py.ingestion.convert_gtfs_rt import GtfsRtConverter, TableData
 from lamp_py.ingestion.converter import ConfigType
 
 from lamp_py.runtime_utils.process_logger import ProcessLogger
-from lamp_py.runtime_utils.remote_files import LAMP, S3Location
+from lamp_py.runtime_utils.remote_files import LAMP, S3_SPRINGBOARD, S3Location
 
 import json
 
 
 # pylint disable=too-many-arguments
-class GtfsRtTripsFullSetConverter(GtfsRtConverter):
+class GtfsRtFullPartitionConverter(GtfsRtConverter):
     """
     Converter that handles GTFS Real Time JSON data
 
@@ -34,31 +34,40 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
         self,
         config_type: ConfigType,
         metadata_queue: Queue[Optional[str]],
-        local_output_location: str,
+        local_output_location: str = "/tmp/gtfs-rt-fullset/",
         remote_output_location: S3Location | None = None,
-        polars_filter: pl.Expr | None = None,  # default to true - which will essentially not filter
+        polars_filter: pl.Expr = pl.lit(True),  # default to true - which will essentially not filter
         max_workers: int = 8,
-        verbose: bool = False,
-        time_chunk_minutes: int = 0,
+        time_chunk_minutes: int = 15,
+        backfill_mode: bool = False,
     ) -> None:
-        GtfsRtConverter.__init__(self, config_type, metadata_queue, max_workers=max_workers, time_chunk_minutes=time_chunk_minutes)
+        GtfsRtConverter.__init__(
+            self, config_type, metadata_queue, max_workers=max_workers, time_chunk_minutes=time_chunk_minutes
+        )
 
+        if time_chunk_minutes < 5:
+            raise ValueError(
+                "time_chunk_minutes must be at least 5 to ensure proper partitioning and avoid too many small files"
+            )
         self.detail = RtTripDetail()
-
         self.tmp_folder = local_output_location
         self.remote_output_location = remote_output_location
-
         self.data_parts: Dict[datetime, TableData] = {}
-
-        self.error_files: List[str] = []
-        self.archive_files: List[str] = []
-        self.table_stats = {}
-
-        self.verbose = verbose
-        self.filter = polars_filter if polars_filter is not None else pl.lit(True)
+        self.filter = polars_filter
+        self.backfill_mode = backfill_mode
 
     def convert(self) -> None:
+        """
+        main convert function - will convert all files in self.files to parquet and
+        upload to s3 if remote_output_location is provided
 
+        Specialization over GtfsRtConverter::convert() - this converter does not apply unique()
+        on the incoming GTFS-RT data, and will partition files by day or by time chunk (e.g. 15 min intervals)
+        depending on the time_chunk_minutes parameter. it will also apply a polars filter to the data at the
+        point of conversion from json.gz to pyarrow table, which should help reduce the amount of data being
+        written out and speed up the conversion process. This converter is applicable for live ingestion and
+        backfill tasks.
+        """
         process_logger = ProcessLogger(
             "fullset_gtfs_parquet_table_creator",
             table_type="gtfs-rt",
@@ -69,16 +78,11 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
 
         table_count = 0
         try:
-            for table, chunk_ts in self.process_files():
+            for table, partition_dt in self.process_files():
                 if table.num_rows == 0:
                     continue
-                partition_dt = chunk_ts
 
-                # include hour/minute in filename when using time-chunk partitioning
-                if self.time_chunk_minutes > 0:
-                    time_suffix = f"{partition_dt.hour:02d}{partition_dt.minute:02d}"
-                else:
-                    time_suffix = ""
+                time_suffix = f"{partition_dt.hour:02d}{partition_dt.minute:02d}"
 
                 path_suffix = os.path.join(
                     LAMP,
@@ -86,7 +90,7 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
                     f"year={partition_dt.year}",
                     f"month={partition_dt.month}",
                     f"day={partition_dt.day}",
-                    f"{partition_dt.isoformat()}{time_suffix}_part_{str(table_count)}.parquet",
+                    f"{partition_dt.isoformat()}{time_suffix}.parquet",
                 )
 
                 local_path = os.path.join(self.tmp_folder, path_suffix)
@@ -94,6 +98,14 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
                 os.makedirs(Path(local_path).parent, exist_ok=True)
 
                 self.write_local_pq_partition(table, local_path)
+
+                # in backfill mode, we don't want to move files around in s3 -
+                # we just want to write the converted files to the output location
+                # (which could be a temp location or the final archive location).
+                # in non-backfill mode, we want to move the original gtfs-rt files from
+                # incoming to archive after successful conversion.
+                if not self.backfill_mode:
+                    self.move_s3_files()
 
                 # mirror on s3 if remote output location is provided
                 if self.remote_output_location is not None:
@@ -109,9 +121,6 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
             process_logger.log_failure(exception)
         else:
             process_logger.log_complete()
-            stats_file = os.path.join(self.tmp_folder, "table_stats.json")
-            with open(stats_file, "a") as f:
-                json.dump(self.table_stats, f, indent=2, default=str)
         finally:
             self.data_parts = {}
 
@@ -119,7 +128,7 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
         """
         just write the file out..
 
-        # this should be roughly sorted by timestamp.
+        # this should be already sorted by timestamp based on how self.files is yielded.
         """
 
         table = pl.from_arrow(table).filter(self.filter).to_arrow()
@@ -131,8 +140,7 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
     def process_files(self) -> Iterable[pyarrow.table]:
         """
         iterate through all of the files to be converted - apply a polars filter to narrow results at the source to reduce write churn - filter at json.gz input level
-
-        only yield a new table when table size crosses over min_rows of yield_check
+        convert to pyarrow tables, group into time chunk intervals, and yield tables for completed intervals as we go
         """
 
         process_logger = ProcessLogger(
@@ -157,22 +165,13 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
                     )
                     continue
 
-                # create key for self.data_parts dictionary
-                # when time_chunk_minutes > 0, partition by sub-day interval;
-                # otherwise partition by day as before
-                if self.time_chunk_minutes > 0:
-                    dt_part = self._interval_key(result_dt)
-                else:
-                    dt_part = datetime(
-                        year=result_dt.year,
-                        month=result_dt.month,
-                        day=result_dt.day,
-                    )
+                # create key for self.data_parts dictionary that bins based on the chunk interval
+                dt_part = self._interval_key(result_dt)
+
                 # create new self.table_groups entry for key if it doesn't exist
                 if dt_part not in self.data_parts:
                     self.data_parts[dt_part] = TableData()
                     self.data_parts[dt_part].table = self.detail.transform_for_write(rt_data)
-
                 else:
                     self.data_parts[dt_part].table = pyarrow.concat_tables(
                         [self.data_parts[dt_part].table, self.detail.transform_for_write(rt_data)]
@@ -180,21 +179,15 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
 
                 self.data_parts[dt_part].files.append(result_filename)
 
-                # we're mapping each gz to a range dt_part. when we have > 1 dt_parts in the map, 
-                # that means we've processed files from at least 2 different time intervals and 
+                # we're mapping each gz to a range dt_part. when we have > 1 dt_parts in the map,
+                # that means we've processed files from at least 2 different time intervals and
                 # can start yielding tables for the intervals that are complete.
-                # this relies on self.files being sorted - enforced by add_files(), 
+                # this relies on self.files being sorted - enforced by add_files(),
                 # and ThreadPoolExecutor/map yields __next__ iterator, i.e. returns in the right order
-                if self.time_chunk_minutes > 0 and len(self.data_parts) > 1:
+                if len(self.data_parts) > 1:
                     yield from self.yield_check_periodic(process_logger, result_dt)
-                # else:
-                #     yield from self.yield_check(process_logger)
 
-        # yield any remaining tables
-        if self.time_chunk_minutes > 0:
-            yield from self.yield_check_periodic(process_logger, flush=True)
-        else:
-            yield from self.yield_check(process_logger, min_rows=-1)
+        yield from self.yield_check_periodic(process_logger, flush=True)
 
         process_logger.add_metadata(file_count=0, number_of_rows=0)
         process_logger.log_complete()
@@ -209,3 +202,67 @@ class GtfsRtTripsFullSetConverter(GtfsRtConverter):
             ("trip_update.trip.direction_id", "ascending"),
             ("trip_update.vehicle.id", "ascending"),
         ]
+
+    def _interval_key(self, ts: datetime) -> datetime:
+        """
+        Truncate a UTC datetime to its wall-clock-aligned interval start.
+
+        For time_chunk_minutes=15:
+            01:07 -> 01:00,  01:15 -> 01:15,  01:29 -> 01:15,  23:59 -> 23:45
+
+        Returns a naive datetime (no timezone) matching the existing data_parts key convention.
+        """
+        total_minutes = ts.hour * 60 + ts.minute
+        aligned_minutes = (total_minutes // self.time_chunk_minutes) * self.time_chunk_minutes
+        return datetime(
+            year=ts.year,
+            month=ts.month,
+            day=ts.day,
+            hour=aligned_minutes // 60,
+            minute=aligned_minutes % 60,
+        )
+
+    def yield_check_periodic(
+        self,
+        process_logger: ProcessLogger,
+        current_ts: datetime = datetime.now(),
+        flush: bool = False,
+    ) -> Iterable[Tuple[pyarrow.table, datetime]]:
+        """
+        Yield completed time-chunk intervals from data_parts.
+
+        When processing files in reverse chronological order, an interval is
+        complete once current_ts is *before* the interval start (we've moved
+        past it going backwards and won't see more data for it).
+
+        When flush=True, yield all remaining intervals regardless of current_ts.
+
+        @current_ts - the feed timestamp of the file just processed
+        @flush - if True, yield everything remaining
+        """
+        current_interval = self._interval_key(current_ts)
+        for iter_ts in list(self.data_parts.keys()):
+            table = self.data_parts[iter_ts].table
+            if table is None:
+                continue
+
+            # yield if we've moved past this interval (reverse order)
+            # or if flushing all remaining data
+            if flush or current_interval < iter_ts:
+
+                # only populate archive_files if we're in live ingestion mode
+                if not self.backfill_mode:
+                    self.archive_files += self.data_parts[iter_ts].files
+
+                process_logger.add_metadata(
+                    file_count=len(self.data_parts[iter_ts].files),
+                    number_of_rows=table.num_rows,
+                    interval_start=iter_ts.isoformat(),
+                )
+                process_logger.log_complete()
+                process_logger.add_metadata(file_count=0, number_of_rows=0, print_log=False)
+                process_logger.log_start()
+
+                yield (table, iter_ts)
+
+                del self.data_parts[iter_ts]
